LayoutTests/svg/dynamic-updates/SVGFEConvolveMatrixElement-svgdom-bias-prop-expected.txt

CONSOLE MESSAGE: line 41: TypeError: 'undefined' is not an object (evaluating 'convolveMatrixElement.bias.baseVal = 0')
FAIL: Timed out waiting for notifyDone to be called
SVG 1.1 dynamic update tests

Tests dynamic updates of the 'bias' property of the SVGFEConvolveMatrixElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL convolveMatrixElement.bias.baseVal should be 0.5. Threw exception TypeError: 'undefined' is not an object (evaluating 'convolveMatrixElement.bias.baseVal')
